# swagger.json
{
    "openapi": "3.0.0",
    "info":{
        "description": "This is the api for users",
        "version":"__version"
    },
    "tags":[
        {
            "name":"PUBLIC ROUTE",
            "description":"Public route"
        }
    ],
    "paths":{
        "/":{
            "summary":"Testing Health Check",
            "get":{
                "description":"Testing health check for the app",
                "tags":["PUBLIC ROUTE"],
                "responses":{
                    "200":{
                        "description":"Health check pass"
                    }
                }
            }
        },
        "/users":{
            "summary":"Testing Health Check",
            "get":{
                "description":"Testing health check for the app",
                "tags":["PUBLIC ROUTE"],
                "parameters":[
                    {
                        "in":"query",
                        "name":"role",
                        "schema":{
                            "type":"string"
                        },
                        "description":"Choose role as admin or user"
                    },
                    {
                        "in":"query",
                        "name":"city",
                        "schema":{
                            "type":"string"
                        },
                        "description":"Enter city"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"Health check pass"
                    }
                }
            }
        },
        "/user/{id}":{
            "summary":"Testing Health Check",
            "get":{
                "description":"Testing health check for the app",
                "tags":["PUBLIC ROUTE"],
                "parameters":[
                    {
                        "in":"path",
                        "name":"id",
                        "required":true,
                        "schema":{
                            "type":"string"
                        },
                        "description":"Enter object id"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"User details"
                    }
                }
            }
        },
        "/addUser":{
            "summary":"Testing Health Check",
            "post":{
                "description":"Testing health check for the app",
                "tags":["PUBLIC ROUTE"],
                "requestBody":{
                    "require":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                "name":"Kiran",
                                "city":"London",
                                "phone":"8965478965",
                                "role":"User",
                                "isActive":true
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"User details"
                    }
                }
            }
        },
        "/updateUser":{
            "summary":"Update user info",
            "put":{
                "description":"Testing user update",
                "tags":["PUBLIC ROUTE"],
                "requestBody":{
                    "require":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                "_id":"",
                                "name":"Kiran",
                                "city":"London",
                                "phone":"8965478965",
                                "role":"User",
                                "isActive":true
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"User details"
                    }
                }
            }
        },
        "/deleteUser":{
            "summary":"Delete user info",
            "delete":{
                "description":"Testing user deletion",
                "tags":["PUBLIC ROUTE"],
                "requestBody":{
                    "require":true,
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "id":"string"
                            },
                            "example":{
                                "_id":""
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"User details"
                    }
                }
            }
        }
    }
}